refactor(llm,core): address PR review feedback
- copy_assets: replace async glob/copy with sync globSync/copyFileSync
- skill: extract skill() factory to its own skill.js / skill.d.ts
- agent: remove section banner comments, drop unnecessary async on execute, simplify typeof guard with Array.includes
- load_content: eliminate duplication between loadContent/findContentDir via shared findContent helper
